docs: update cli reference
sync up cli references
